Krishna_Kumar_Singh-Data_Flow_Pipeline-Task-Zip_Contains_Screenshot_And_Code/Geolocation/GeoLocation/requirements.txt
backoff==1.10.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
kafka-python==2.0.2
opencage==1.2.2
pip==21.0.1
pycparser==2.20
pygeohash==1.2.0
pyOpenSSL==20.0.1
requests==2.25.1
setuptools==53.0.0
six==1.15.0
urllib3==1.26.3